move fetch coords from online db to new component
