+++ baby-gru/src/components/navbar-menus/MoorhenFileMenu.tsx
@@ -1,7 +1,7 @@
-import { Form, Button, InputGroup, SplitButton, Dropdown, Stack } from "react-bootstrap";
+import { Form, Stack } from "react-bootstrap";
 import { MoorhenMolecule } from "../../utils/MoorhenMolecule";
-import { MoorhenMap } from "../../utils/MoorhenMap";
-import { useState, useRef, useCallback } from "react";
+import { useState, useCallback } from "react";
+import { MoorhenFetchOnlineSourcesForm } from "../form/MoorhenFetchOnlineSourcesForm"
 import { MoorhenLoadTutorialDataMenuItem } from "../menu-item/MoorhenLoadTutorialDataMenuItem"
 import { MoorhenAssociateReflectionsToMap } from "../menu-item/MoorhenAssociateReflectionsToMap";
 import { MoorhenAutoOpenMtzMenuItem } from "../menu-item/MoorhenAutoOpenMtzMenuItem"
@@ -12,23 +12,18 @@ import { MoorhenImportMapCoefficientsMenuItem } from "../menu-item/MoorhenImport
 import { MoorhenDeleteEverythingMenuItem } from "../menu-item/MoorhenDeleteEverythingMenuItem"
 import { IconButton, MenuItem } from "@mui/material";
 import { RadioButtonCheckedOutlined, StopCircleOutlined, WarningOutlined } from "@mui/icons-material";
-import { convertViewtoPx, doDownload, readTextFile, getMultiColourRuleArgs, loadSessionData, guid } from "../../utils/MoorhenUtils";
+import { convertViewtoPx, doDownload, readTextFile, loadSessionData, guid } from "../../utils/MoorhenUtils";
 import { getBackupLabel } from "../../utils/MoorhenTimeCapsule"
 import { MoorhenNavBarExtendedControlsInterface } from "./MoorhenNavBar";
 import { MoorhenNotification } from "../misc/MoorhenNotification";
 import { moorhen } from "../../types/moorhen";
-import { useSelector, useDispatch, batch } from 'react-redux';
-import { setActiveMap, setNotificationContent } from "../../store/generalStatesSlice";
-import { addMolecule, addMoleculeList } from "../../store/moleculesSlice";
-import { addMap } from "../../store/mapsSlice";
+import { useSelector, useDispatch } from 'react-redux';
+import { setNotificationContent } from "../../store/generalStatesSlice";
+import { addMoleculeList } from "../../store/moleculesSlice";
 
 export const MoorhenFileMenu = (props: MoorhenNavBarExtendedControlsInterface) => {
-    const pdbCodeFetchInputRef = useRef<HTMLInputElement | null>(null);
-    const fetchMapDataCheckRef = useRef<HTMLInputElement | null>(null);
-
+    
     const [popoverIsShown, setPopoverIsShown] = useState<boolean>(false)
-    const [remoteSource, setRemoteSource] = useState<string>("PDBe")
-    const [isValidPdbId, setIsValidPdbId] = useState<boolean>(true)
     
     const dispatch = useDispatch()
     const maps = useSelector((state: moorhen.State) => state.maps)
@@ -38,7 +33,7 @@ export const MoorhenFileMenu = (props: MoorhenNavBarExtendedControlsInterface) =
     const backgroundColor = useSelector((state: moorhen.State) => state.canvasStates.backgroundColor)
     const molecules = useSelector((state: moorhen.State) => state.molecules)
 
-    const { commandCentre, glRef, monomerLibraryPath } = props;
+    const { commandCentre, glRef, monomerLibraryPath, setBusy } = props;
 
     const getWarningToast = (message: string) => <MoorhenNotification key={guid()} hideDelay={3000} width={20}>
             <><WarningOutlined style={{margin: 0}}/>
@@ -83,147 +78,6 @@ export const MoorhenFileMenu = (props: MoorhenNavBarExtendedControlsInterface) =
         newMolecule.defaultBondOptions.smoothness = defaultBondSmoothness
         await newMolecule.loadToCootFromFile(file)        
         return newMolecule        
-    }
-
-    const fetchFiles = (): void => {
-        props.setBusy(true)
-        if (remoteSource === "PDBe") {
-            fetchFilesFromEBI()
-        } else if (remoteSource === "PDB-REDO") {
-            fetchFilesFromPDBRedo()
-        } else if (remoteSource === 'AFDB') {
-            fetchFilesFromAFDB()
-        } else if (remoteSource === 'EMDB') {
-            fetchMapFromEMDB()
-        } else {
-            console.log(`Unrecognised remote source! ${remoteSource}`)
-        }
-    }
-
-    const fetchMapFromEMDB = () => {
-        const emdbCode = pdbCodeFetchInputRef.current.value.toLowerCase()
-        if (emdbCode) {
-            const mapUrl = `https://ftp.ebi.ac.uk/pub/databases/emdb/structures/EMD-${emdbCode}/map/emd_${emdbCode}.map.gz`
-            fetchMapFromURL(mapUrl, `${emdbCode}.map.gz`, false)
-            .then(newMap => newMap.centreOnMap())
-        } else {
-            console.log('Error: no EMDB entry provided')
-        }
-    }
-
-    const fetchFilesFromCOD = () => {
-        const entryCod: string = pdbCodeFetchInputRef.current.value.toLowerCase()
-        const codUrl = `http://www.crystallography.net/cod/${entryCod}.cif`
-        if (entryCod) {
-            Promise.all([
-                fetchMoleculeFromURL(codUrl, `cod-${entryCod}`),
-            ])
-        } else {
-            console.log('Error: no COD entry')
-        }
-    }
-
-    const fetchFilesFromEBI = () => {
-        const pdbCode = pdbCodeFetchInputRef.current.value.toLowerCase()
-        const coordUrl = `https://www.ebi.ac.uk/pdbe/entry-files/download/${pdbCode}.cif`
-        const mapUrl = `https://www.ebi.ac.uk/pdbe/entry-files/${pdbCode}.ccp4`
-        const diffMapUrl = `https://www.ebi.ac.uk/pdbe/entry-files/${pdbCode}_diff.ccp4`
-        if (pdbCode && fetchMapDataCheckRef.current.checked) {
-            Promise.all([
-                fetchMoleculeFromURL(coordUrl, pdbCode),
-                fetchMapFromURL(mapUrl, `${pdbCode}-map`),
-                fetchMapFromURL(diffMapUrl, `${pdbCode}-map`, true)
-            ])
-        } else if (pdbCode) {
-            fetchMoleculeFromURL(coordUrl, pdbCode)
-        }
-    }
-
-    const fetchFilesFromAFDB = () => {
-        const uniprotID: string = pdbCodeFetchInputRef.current.value.toUpperCase()
-        const coordUrl = `https://alphafold.ebi.ac.uk/files/AF-${uniprotID}-F1-model_v4.pdb`
-        if (uniprotID) {
-            fetchMoleculeFromURL(coordUrl, `${uniprotID}`, true)
-        }
-    }
-
-    const fetchFilesFromPDBRedo = () => {
-        const pdbCode = pdbCodeFetchInputRef.current.value
-        const coordUrl = `https://pdb-redo.eu/db/${pdbCode}/${pdbCode}_final.pdb`
-        const mtzUrl = `https://pdb-redo.eu/db/${pdbCode}/${pdbCode}_final.mtz/`
-        if (pdbCode && fetchMapDataCheckRef.current.checked) {
-            Promise.all([
-                fetchMoleculeFromURL(coordUrl, `${pdbCode}-redo`),
-                fetchMtzFromURL(mtzUrl, `${pdbCode}-map-redo`,  {F: "FWT", PHI: "PHWT", Fobs: 'FP', SigFobs: 'SIGFP', FreeR: 'FREE', isDifference: false, useWeight: false, calcStructFact: true}),  
-                fetchMtzFromURL(mtzUrl, `${pdbCode}-map-redo`,  {F: "DELFWT", PHI: "PHDELWT", isDifference: true, useWeight: false})    
-            ])
-        } else if (pdbCode) {
-            fetchMoleculeFromURL(coordUrl, `${pdbCode}-redo`)
-        }
-    }
-
-    const fetchMoleculeFromURL = async (url: RequestInfo | URL, molName: string, isAF2?: boolean): Promise<moorhen.Molecule> => {
-        const newMolecule = new MoorhenMolecule(commandCentre, glRef, monomerLibraryPath)
-        newMolecule.setBackgroundColour(backgroundColor)
-        newMolecule.defaultBondOptions.smoothness = defaultBondSmoothness
-        try {
-            await newMolecule.loadToCootFromURL(url, molName)
-            if (newMolecule.molNo === -1) {
-                throw new Error("Cannot read the fetched molecule...")
-            } else if (isAF2) {
-                const newRule = {
-                    args: [getMultiColourRuleArgs(newMolecule, 'af2-plddt')],
-                    isMultiColourRule: true,
-                    ruleType: 'af2-plddt',
-                    label: `//*`
-                }
-                newMolecule.defaultColourRules = [newRule]
-            }
-            await newMolecule.fetchIfDirtyAndDraw(newMolecule.atomCount >= 50000 ? 'CRs' : 'CBs')
-            await newMolecule.centreOn('/*/*/*/*', true)
-            dispatch( addMolecule(newMolecule) )
-            return newMolecule
-        } catch (err) {
-            dispatch(setNotificationContent(getWarningToast(`Failed to read molecule`)))
-            console.log(`Cannot fetch molecule from ${url}`)
-            setIsValidPdbId(false)
-            props.setBusy(false)
-        }
-    }
-
-    const fetchMapFromURL = async (url: RequestInfo | URL, mapName: string, isDiffMap: boolean = false): Promise<moorhen.Map> => {
-        const newMap = new MoorhenMap(commandCentre, glRef)
-        try {
-            await newMap.loadToCootFromMapURL(url, mapName, isDiffMap)
-            if (newMap.molNo === -1) throw new Error("Cannot read the fetched map...")
-            batch(() => {
-                dispatch( addMap(newMap) )
-                dispatch( setActiveMap(newMap) )
-            })
-        } catch (err) {
-            console.warn(err)
-            dispatch(setNotificationContent(getWarningToast(`Failed to read map`)))
-            console.log(`Cannot fetch map from ${url}`)
-            props.setBusy(false)
-        }
-        return newMap
-    }
-
-    const fetchMtzFromURL = async (url: RequestInfo | URL, mapName: string, selectedColumns: moorhen.selectedMtzColumns): Promise<moorhen.Map> => {
-        const newMap = new MoorhenMap(commandCentre, glRef)
-        try {
-            await newMap.loadToCootFromMtzURL(url, mapName, selectedColumns)
-            if (newMap.molNo === -1) throw new Error("Cannot read the fetched mtz...")
-            batch(() => {
-                dispatch( addMap(newMap) )
-                dispatch( setActiveMap(newMap) )
-            })
-        } catch {
-            dispatch(setNotificationContent(getWarningToast(`Failed to read mtz`)))
-            console.log(`Cannot fetch mtz from ${url}`)
-            props.setBusy(false)
-        }
-        return newMap
     }
 
     const handleSessionUpload = async (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -316,49 +170,20 @@ export const MoorhenFileMenu = (props: MoorhenNavBarExtendedControlsInterface) =
 
     return <>
                 <div style={{maxHeight: convertViewtoPx(65, height), overflowY: 'auto'}}>
+                    
                     {!props.disableFileUploads && 
                     <Form.Group className='moorhen-form-group' controlId="upload-coordinates-form">
                         <Form.Label>Coordinates</Form.Label>
                         <Form.Control type="file" accept=".pdb, .mmcif, .cif, .ent" multiple={true} onChange={(e: React.ChangeEvent<HTMLInputElement>) => { loadPdbFiles(e.target.files) }}/>
-                    </Form.Group>
-                    }
-                    <Form.Group className='moorhen-form-group' controlId="fetch-pdbe-form">
-                        <Form.Label>Fetch coords from online services</Form.Label>
-                        <InputGroup>
-                            <SplitButton title={remoteSource} id="fetch-coords-online-source-select">
-                                {/* @ts-ignore */}
-                                <Dropdown.Item key="PDBe" href="#" onClick={() => {
-                                    setRemoteSource("PDBe")
-                                }}>PDBe</Dropdown.Item>
-                                <Dropdown.Item key="PDB-REDO" href="#" onClick={() => {
-                                    setRemoteSource("PDB-REDO")
-                                }}>PDB-REDO</Dropdown.Item>
-                                <Dropdown.Item key="AFDB" href="#" onClick={() => {
-                                    setRemoteSource("AFDB")
-                                }}>AlphaFold DB</Dropdown.Item>
-                                <Dropdown.Item key="EMDB" href="#" onClick={() => {
-                                    setRemoteSource("EMDB")
-                                }}>EMDB</Dropdown.Item>
-                            </SplitButton>
-                            <Form.Control type="text" style={{borderColor: isValidPdbId ? '' : 'red'}}  ref={pdbCodeFetchInputRef} onKeyDown={(e) => {
-                                setIsValidPdbId(true)
-                                if (e.code === 'Enter') {
-                                    fetchFiles()
-                                }
-                            }}/>
-                            <Button variant="light" onClick={fetchFiles}>
-                                Fetch
-                            </Button>
-                        </InputGroup>
-                        <Form.Label style={{display: isValidPdbId ? 'none' : 'block', alignContent: 'center' ,textAlign: 'center'}}>Problem fetching</Form.Label>
-                        <Form.Check style={{ marginTop: '0.5rem' }} ref={fetchMapDataCheckRef} label={'fetch data for map'} name={`fetchMapData`} type="checkbox" />
-                    </Form.Group>
+                    </Form.Group>}
+                    
+                    <MoorhenFetchOnlineSourcesForm commandCentre={commandCentre} glRef={glRef} setBusy={setBusy} monomerLibraryPath={monomerLibraryPath} />
+                    
                     {!props.disableFileUploads && 
                     <Form.Group className='moorhen-form-group' controlId="upload-session-form">
                         <Form.Label>Load from stored session</Form.Label>
                         <Form.Control type="file" accept=".json" multiple={false} onChange={handleSessionUpload}/>
-                    </Form.Group>
-                    }
+                    </Form.Group>}
                     
                     <hr></hr>
 
@@ -375,8 +200,7 @@ export const MoorhenFileMenu = (props: MoorhenNavBarExtendedControlsInterface) =
                         <MoorhenAutoOpenMtzMenuItem {...menuItemProps} />
                         <MoorhenImportMapCoefficientsMenuItem {...menuItemProps} />
                         <MoorhenImportMapMenuItem {...menuItemProps} />
-                    </>
-                    }
+                    </>}
                     
                     <MoorhenImportFSigFMenuItem {...menuItemProps} />
